iot-simulator/requirements.txt
paho-mqtt
numpy
python-dotenv
